Move CLI stuff to its own file
